requirements.txt
asgiref==3.8.1
boto3==1.34.90
botocore==1.34.90
dj-database-url==2.1.0
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
drf-nested-routers==0.93.5
jmespath==1.0.1
pillow==10.3.0
psycopg2==2.9.9
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2023.3.post1
s3transfer==0.10.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.11.0
tzdata==2023.4
urllib3==2.2.1
whitenoise==6.6.0
gunicorn==23.0.0
requests
aiohttp